LPS-93019 portal-search: Move relevant logic to SearchBarPortletSharedSearchContributor and revert the rest of the changes
https://issues.liferay.com/browse/LPS-93019
